+++ sdk/program/src/message/sanitized.rs
@@ -451,21 +451,7 @@ impl TransactionSignatureDetails {
 
 #[cfg(test)]
 mod tests {
-    use {
-        super::*,
-        crate::message::v0,
-        solana_account_info::AccountInfo,
-        solana_instruction::{AccountMeta, BorrowedAccountMeta, BorrowedInstruction, Instruction},
-        solana_program_error::ProgramError,
-        solana_sanitize::SanitizeError,
-        solana_sdk_ids::sysvar::instructions::id,
-        solana_sysvar::instructions::{
-            construct_instructions_data, deserialize_instruction, get_instruction_relative,
-            load_current_index_checked, load_instruction_at_checked, serialize_instructions,
-            store_current_index,
-        },
-        std::collections::HashSet,
-    };
+    use {super::*, crate::message::v0, std::collections::HashSet};
 
     #[test]
     fn test_try_from_legacy_message() {
@@ -716,346 +702,5 @@ mod tests {
         assert_eq!(2, signature_details.num_secp256k1_instruction_signatures);
         // expect 5 ed25519 instruction signatures from mock_ed25519_instr
         assert_eq!(5, signature_details.num_ed25519_instruction_signatures);
-    }
-
-    #[derive(Copy, Clone)]
-    struct MakeInstructionParams {
-        program_id: Pubkey,
-        account_key: Pubkey,
-        is_signer: bool,
-        is_writable: bool,
-    }
-
-    fn make_borrowed_instruction<'a>(params: &'a MakeInstructionParams) -> BorrowedInstruction<'a> {
-        let MakeInstructionParams {
-            program_id,
-            account_key,
-            is_signer,
-            is_writable,
-        } = params;
-        BorrowedInstruction {
-            program_id,
-            accounts: vec![BorrowedAccountMeta {
-                pubkey: account_key,
-                is_signer: *is_signer,
-                is_writable: *is_writable,
-            }],
-            data: &[0],
-        }
-    }
-
-    fn make_instruction(params: MakeInstructionParams) -> Instruction {
-        let MakeInstructionParams {
-            program_id,
-            account_key,
-            is_signer,
-            is_writable,
-        } = params;
-        Instruction {
-            program_id,
-            accounts: vec![AccountMeta {
-                pubkey: account_key,
-                is_signer,
-                is_writable,
-            }],
-            data: vec![0],
-        }
-    }
-
-    #[test]
-    fn test_load_instruction_at_checked() {
-        let program_id0 = Pubkey::new_unique();
-        let program_id1 = Pubkey::new_unique();
-        let account_key0 = Pubkey::new_unique();
-        let account_key1 = Pubkey::new_unique();
-        let params0 = MakeInstructionParams {
-            program_id: program_id0,
-            account_key: account_key0,
-            is_signer: false,
-            is_writable: false,
-        };
-        let params1 = MakeInstructionParams {
-            program_id: program_id1,
-            account_key: account_key1,
-            is_signer: false,
-            is_writable: false,
-        };
-        let instruction0 = make_instruction(params0);
-        let instruction1 = make_instruction(params1);
-        let borrowed_instruction0 = make_borrowed_instruction(&params0);
-        let borrowed_instruction1 = make_borrowed_instruction(&params1);
-        let key = id();
-        let mut lamports = 0;
-        let mut data = construct_instructions_data(&[borrowed_instruction0, borrowed_instruction1]);
-        let owner = solana_sdk_ids::sysvar::id();
-        let mut account_info = AccountInfo::new(
-            &key,
-            false,
-            false,
-            &mut lamports,
-            &mut data,
-            &owner,
-            false,
-            0,
-        );
-
-        assert_eq!(
-            instruction0,
-            load_instruction_at_checked(0, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction1,
-            load_instruction_at_checked(1, &account_info).unwrap()
-        );
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            load_instruction_at_checked(2, &account_info)
-        );
-
-        let key = Pubkey::new_unique();
-        account_info.key = &key;
-        assert_eq!(
-            Err(ProgramError::UnsupportedSysvar),
-            load_instruction_at_checked(2, &account_info)
-        );
-    }
-
-    #[test]
-    fn test_load_current_index_checked() {
-        let program_id0 = Pubkey::new_unique();
-        let program_id1 = Pubkey::new_unique();
-        let account_key0 = Pubkey::new_unique();
-        let account_key1 = Pubkey::new_unique();
-        let params0 = MakeInstructionParams {
-            program_id: program_id0,
-            account_key: account_key0,
-            is_signer: false,
-            is_writable: false,
-        };
-        let params1 = MakeInstructionParams {
-            program_id: program_id1,
-            account_key: account_key1,
-            is_signer: false,
-            is_writable: false,
-        };
-        let borrowed_instruction0 = make_borrowed_instruction(&params0);
-        let borrowed_instruction1 = make_borrowed_instruction(&params1);
-
-        let key = id();
-        let mut lamports = 0;
-        let mut data = construct_instructions_data(&[borrowed_instruction0, borrowed_instruction1]);
-        store_current_index(&mut data, 1);
-        let owner = solana_sdk_ids::sysvar::id();
-        let mut account_info = AccountInfo::new(
-            &key,
-            false,
-            false,
-            &mut lamports,
-            &mut data,
-            &owner,
-            false,
-            0,
-        );
-
-        assert_eq!(1, load_current_index_checked(&account_info).unwrap());
-        {
-            let mut data = account_info.try_borrow_mut_data().unwrap();
-            store_current_index(&mut data, 0);
-        }
-        assert_eq!(0, load_current_index_checked(&account_info).unwrap());
-
-        let key = Pubkey::new_unique();
-        account_info.key = &key;
-        assert_eq!(
-            Err(ProgramError::UnsupportedSysvar),
-            load_current_index_checked(&account_info)
-        );
-    }
-
-    #[test]
-    fn test_get_instruction_relative() {
-        let program_id0 = Pubkey::new_unique();
-        let program_id1 = Pubkey::new_unique();
-        let program_id2 = Pubkey::new_unique();
-        let account_key0 = Pubkey::new_unique();
-        let account_key1 = Pubkey::new_unique();
-        let account_key2 = Pubkey::new_unique();
-        let params0 = MakeInstructionParams {
-            program_id: program_id0,
-            account_key: account_key0,
-            is_signer: false,
-            is_writable: false,
-        };
-        let params1 = MakeInstructionParams {
-            program_id: program_id1,
-            account_key: account_key1,
-            is_signer: false,
-            is_writable: false,
-        };
-        let params2 = MakeInstructionParams {
-            program_id: program_id2,
-            account_key: account_key2,
-            is_signer: false,
-            is_writable: false,
-        };
-        let instruction0 = make_instruction(params0);
-        let instruction1 = make_instruction(params1);
-        let instruction2 = make_instruction(params2);
-        let borrowed_instruction0 = make_borrowed_instruction(&params0);
-        let borrowed_instruction1 = make_borrowed_instruction(&params1);
-        let borrowed_instruction2 = make_borrowed_instruction(&params2);
-
-        let key = id();
-        let mut lamports = 0;
-        let mut data = construct_instructions_data(&[
-            borrowed_instruction0,
-            borrowed_instruction1,
-            borrowed_instruction2,
-        ]);
-        store_current_index(&mut data, 1);
-        let owner = solana_sdk_ids::sysvar::id();
-        let mut account_info = AccountInfo::new(
-            &key,
-            false,
-            false,
-            &mut lamports,
-            &mut data,
-            &owner,
-            false,
-            0,
-        );
-
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(-2, &account_info)
-        );
-        assert_eq!(
-            instruction0,
-            get_instruction_relative(-1, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction1,
-            get_instruction_relative(0, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction2,
-            get_instruction_relative(1, &account_info).unwrap()
-        );
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(2, &account_info)
-        );
-        {
-            let mut data = account_info.try_borrow_mut_data().unwrap();
-            store_current_index(&mut data, 0);
-        }
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(-1, &account_info)
-        );
-        assert_eq!(
-            instruction0,
-            get_instruction_relative(0, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction1,
-            get_instruction_relative(1, &account_info).unwrap()
-        );
-        assert_eq!(
-            instruction2,
-            get_instruction_relative(2, &account_info).unwrap()
-        );
-        assert_eq!(
-            Err(ProgramError::InvalidArgument),
-            get_instruction_relative(3, &account_info)
-        );
-
-        let key = Pubkey::new_unique();
-        account_info.key = &key;
-        assert_eq!(
-            Err(ProgramError::UnsupportedSysvar),
-            get_instruction_relative(0, &account_info)
-        );
-    }
-
-    #[test]
-    fn test_serialize_instructions() {
-        let program_id0 = Pubkey::new_unique();
-        let program_id1 = Pubkey::new_unique();
-        let id0 = Pubkey::new_unique();
-        let id1 = Pubkey::new_unique();
-        let id2 = Pubkey::new_unique();
-        let id3 = Pubkey::new_unique();
-        let params = vec![
-            MakeInstructionParams {
-                program_id: program_id0,
-                account_key: id0,
-                is_signer: false,
-                is_writable: true,
-            },
-            MakeInstructionParams {
-                program_id: program_id0,
-                account_key: id1,
-                is_signer: true,
-                is_writable: true,
-            },
-            MakeInstructionParams {
-                program_id: program_id1,
-                account_key: id2,
-                is_signer: false,
-                is_writable: false,
-            },
-            MakeInstructionParams {
-                program_id: program_id1,
-                account_key: id3,
-                is_signer: true,
-                is_writable: false,
-            },
-        ];
-        let instructions: Vec<Instruction> =
-            params.clone().into_iter().map(make_instruction).collect();
-        let borrowed_instructions: Vec<BorrowedInstruction> =
-            params.iter().map(make_borrowed_instruction).collect();
-
-        let serialized = serialize_instructions(&borrowed_instructions);
-
-        // assert that deserialize_instruction is compatible with SanitizedMessage::serialize_instructions
-        for (i, instruction) in instructions.iter().enumerate() {
-            assert_eq!(
-                deserialize_instruction(i, &serialized).unwrap(),
-                *instruction
-            );
-        }
-    }
-
-    #[test]
-    fn test_decompile_instructions_out_of_bounds() {
-        let program_id0 = Pubkey::new_unique();
-        let id0 = Pubkey::new_unique();
-        let id1 = Pubkey::new_unique();
-        let params = vec![
-            MakeInstructionParams {
-                program_id: program_id0,
-                account_key: id0,
-                is_signer: false,
-                is_writable: true,
-            },
-            MakeInstructionParams {
-                program_id: program_id0,
-                account_key: id1,
-                is_signer: true,
-                is_writable: true,
-            },
-        ];
-        let instructions: Vec<Instruction> =
-            params.clone().into_iter().map(make_instruction).collect();
-        let borrowed_instructions: Vec<BorrowedInstruction> =
-            params.iter().map(make_borrowed_instruction).collect();
-
-        let serialized = serialize_instructions(&borrowed_instructions);
-        assert_eq!(
-            deserialize_instruction(instructions.len(), &serialized).unwrap_err(),
-            SanitizeError::IndexOutOfBounds,
-        );
     }
 }
